landmarks/requirements.txt
numpy
pandas
matplotlib
scipy
torch
torchfile==0.1.0
torchvision
Pillow
tqdm